优化加载

+++ app/src/main/java/com/soya/launcher/cache/AppCache.kt
@@ -8,7 +8,7 @@ object AppCache {
     var updateInteval:String by serial(default = "day")
     var lastTipTime:Long by serial(default = 0L)
     var updateInfo:String by serial()
-    var reqId:Long by serial(0L)
+    var reqId:Int by serial(0)
     var homeData: HomeDataList by serial(default = HomeDataList(mutableMapOf()))
 
 }
